requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
python-dotenv>=0.19.0
google-generativeai>=0.8.0
tavily-python>=0.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0 